fixed bug with empty lists
