+++ packages/netsuite-adapter/src/change_validators/saved_search_modifications.ts
@@ -1,0 +1,35 @@
+/*
+*                      Copyright 2021 Salto Labs Ltd.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with
+* the License.  You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+import {
+  isInstanceElement, ChangeValidator, getChangeElement, isModificationChange,
+} from '@salto-io/adapter-api'
+import { SAVED_SEARCH } from '../constants'
+
+const changeValidator: ChangeValidator = async changes => (
+  changes
+    .filter(isModificationChange)
+    .map(getChangeElement)
+    .filter(elem => isInstanceElement(elem))
+    .filter(elem => elem.elemID.typeName === SAVED_SEARCH)
+    .map(({ elemID }) => ({
+      elemID,
+      severity: 'Error',
+      message: 'Type definitions are read only',
+      detailedMessage: `Changing (${elemID.name}) is not supported`,
+    }))
+)
+
+export default changeValidator
